Adding prop_map to variables sent to report_dataset
